Put encryption_properties.h/cc content in encryption.h/cc
